refactor: Move files to src directory
