+++ projects/llm-wrapper-api/.vscode/launch.json
@@ -6,7 +6,7 @@
       "type": "debugpy",
       "request": "launch",
       "module": "uvicorn",
-      "args": ["app.main:app", "--reload", "--host", "0.0.0.0"],
+      "args": ["src.main:app", "--reload", "--host", "0.0.0.0"],
       "console": "integratedTerminal",
       "justMyCode": true
     },
