+++ projects/llm-wrapper-api/Dockerfile
@@ -3,19 +3,21 @@ FROM python:3.13-alpine AS base
 ARG USERNAME=pyuser
 ARG COMMIT_HASH=""
 
-# -----------------------------------------------
-# builder stage
-# -----------------------------------------------
-FROM base AS builder
-COPY --from=ghcr.io/astral-sh/uv:latest /uv /uvx /bin/
-
 WORKDIR /app
 COPY . .
+
+# -----------------------------------------------
+# production builder stage
+# -----------------------------------------------
+FROM base AS builder
+
+COPY --from=ghcr.io/astral-sh/uv:latest /uv /uvx /bin/
+COPY --from=base /app .
 
 RUN uv sync --frozen --no-dev
 
 # -----------------------------------------------
-# runner stage
+# production runner stage
 # -----------------------------------------------
 FROM base AS runner
 
@@ -26,14 +28,12 @@ ENV TZ=Asia/Tokyo
 ENV PATH="/app/.venv/bin:$PATH"
 ENV APP_VERSION=${COMMIT_HASH}
 
-WORKDIR /app
-
 RUN adduser -D $USERNAME
 
 COPY --from=builder --chown=$USERNAME:$USERNAME /app/.venv ./.venv
-COPY --from=builder --chown=$USERNAME:$USERNAME /app/app ./app
+COPY --from=builder --chown=$USERNAME:$USERNAME /app/src ./src
 
 USER $USERNAME
 
 EXPOSE 8000
-CMD ["python", ".venv/bin/uvicorn", "app.main:app", "--host", "0.0.0.0"]
+CMD ["python", ".venv/bin/uvicorn", "src.main:app", "--host", "0.0.0.0"]
