+++ projects/llm-wrapper-api/tests/test_main.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from fastapi.testclient import TestClient
 
-from app.main import app
+from src.main import app
 
 
 @pytest.fixture
